2.3.0
Added images for the types camera, mail server, nfs server, web server,
and unknown. The page now shows more information in the limited box for
each computer: right now, it shows 1) the number of users online on that
computer, and 2) when there's either "no paper" or "low ink" (< 5%) in a
printer. The amount of uptime / downtime is now displayed on the layout
next to the "Up" / "Down" word, in a very limited form.

Re-wrote the date time display function to support a "short" and "long"
time-ago format (for example "1y" or "1 year, 2 days, 3 hours ago"). The
short format is used inside the boxes and the long format is used with
all date-times shown in the tooltips.

Tooltips now calculate the "sessions" object when new host data arrives
instead of every time a tooltip is (re-)created. This allowed us to show
the number of unique users on the page.

Tooltips now show the associated icon in the header line.