Now supports multiple CPUs and swap memory.
Added support for multiple CPUs on a single computer. Will now correctly
handle the display of complicated (hyper-)threading setups, if present.

Added support for showing current RAM usage on a computer, with a green
bar in the UI. Added support for showing the current swap usage and
total swap space. SunOS systems (niagara1) do not support this yet.

Changed the threshold for the "near full" hard disk to 90% or more.
Made a threshold for the "near empty" printer ink at 10% or less.
Made a threshold for the "near full" RAM at 90% or more.
If those thresholds are passed, the bars will appear red instead.